requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.3
django-filter==23.5
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
mongoengine==0.28.2
